+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### 1.64 (2018-04-16)
+### 1.65 (2018-04-16)
   * Bugfix: #534 VersionStore: overwriting a symbol with different dtype (but same data format) does not
                  raise exceptions anymore
   * Bugfix: #531 arctic_prune_versions: clean broken snapshot references before pruning